web_service/requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
pinecone-client==5.0.1
langchain==0.3.13
langchain-groq==0.2.2
langchain-pinecone==0.2.0
langchain-core==0.3.28
PyMuPDF==1.25.1
pillow==11.0.0
requests==2.32.3
cloudinary==1.41.0